Object detection live demo (#290)
* Added readmes

* Moved video player to utils

* Added images

* Added object detection live demo

* Replaced static image with gif

* Converted notebook_utils with jupyter command

* Added keywords to arguments

* Small documentation changes

* Changed threshold a little

* Apply suggestions from code review


* Improved notebook according to the review
